+++ gameplay/src/Vector2.cpp
@@ -1,7 +1,3 @@
-/*
- * Vector2.cpp
- */
-
 #include "Base.h"
 #include "Vector2.h"
 
@@ -133,7 +129,7 @@ float Vector2::distance(const Vector2& v) const
     float dx = v.x - x;
     float dy = v.y - y;
 
-    return sqrtf(dx * dx + dy * dy);
+    return sqrt(dx * dx + dy * dy);
 }
 
 float Vector2::distanceSquared(const Vector2& v) const
@@ -155,7 +151,7 @@ float Vector2::dot(const Vector2& v1, const Vector2& v2)
 
 float Vector2::length() const
 {
-    return sqrtf(x * x + y * y);
+    return sqrt(x * x + y * y);
 }
 
 float Vector2::lengthSquared() const
@@ -189,7 +185,7 @@ void Vector2::normalize(Vector2* dst)
     if (n == 1.0f)
         return;
 
-    n = sqrtf(n);
+    n = sqrt(n);
     // Too close to zero.
     if (n < MATH_TOLERANCE)
         return;
